fix(coreos-base/oem-vagrant): Add coreos-setup-environment
In vagrant /etc/environment is set up by cloudinit but other services
still depend on the coreos-setup-environment.service unit as a signal
that /etc/environment is valid. Also the OEM needs to provide this
script in order to block the likely to be broken defaults from being
written. This just provides a dumb script that blocks.

Use system-cloudinit@.service because it does not depend on
coreos-setup-environment.service. Otherwise we would dead lock on a
circular dependency.